registration successful

+++ src/image_pipeline.cpp
@@ -2,7 +2,7 @@
 
 namespace  image_pipeline {
 
-  Pose::Pose(const std::string& frame_id, const Eigen::Affine3f transform)
+  Pose::Pose(const std::string& frame_id, const Eigen::Affine3d transform)
     :
       frame_id(frame_id),
       transform(transform)
